+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added sample output for `enable_verbose_debug_messages(style='rainbow')` to
+  [docs](https://id-translation.readthedocs.io/en/latest/_static/logging-style-rainbow.html).
+
 ### Fixed
 - The `MultiFetcher` will no longer warn twice for required children without sources.
 - Fix indentation in `enable_verbose_debug_messages()` messages when using `style='rainbow`.